Rename hardwareConfig to config
