+++ src/managers/taskmanager.cpp
@@ -13,7 +13,7 @@
 #include "managers/sensormanager.h"
 #include "screens/lockscreen.h"
 #include "screens/mainscreen.h"
-#include "utils/hardwareConfig.h"
+#include "utils/config.h"
 #include "utils/timeUtils.h"
 
 TaskManager *TaskManager::s_instance = nullptr;
